requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arxiv==2.3.1
atpublic==6.0.2
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
brotli==1.2.0
bs4==0.0.2
build==1.3.0
CacheControl==0.14.4
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.2
chromadb==1.3.5
click==8.3.1
cloudscraper==1.2.71
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dashscope==1.25.2
dataclasses-json==0.6.7
ddgs==9.9.2
distro==1.9.0
dotenv==0.9.9
duckduckgo_search==8.1.1
durationpy==0.10
extra-streamlit-components==0.1.81
fake-useragent==2.2.0
fastapi==0.124.4
feedparser==6.0.12
filelock==3.20.0
firebase_admin==7.1.0
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsrs==6.3.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.16.1
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-deepseek==1.0.1
langchain-experimental==0.4.0
langchain-huggingface==1.1.0
langchain-openai==1.1.0
langchain-tavily==0.2.13
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.48
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mcp==1.25.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip_system_certs==5.3
pluggy==1.6.0
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.5
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==6.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.21
python-pptx==1.0.2
python-telegram-bot==22.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
searchx==1.0.1
selenium==4.26.1
sentence-transformers==5.1.2
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.4
starlette==0.50.0
streamlit==1.51.0
streamlit-paste-button==0.1.2
sympy==1.14.0
tavily-python==0.7.13
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
trio==0.31.0
trio-websocket==0.12.2
triton==3.5.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wikipedia==1.4.0
world==5.2.1
wsproto==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.12.8
zipp==3.23.0
zstandard==0.25.0
